requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
python-multipart==0.0.6
python-dateutil==2.8.2
bleak
pyyaml
